Cleaned up code
git-svn-id: https://svn.csail.mit.edu/streamit/branches/streamit-3.0@11542 fb122d4e-f3b4-4596-99d2-990b948fa53d

+++ streams/src/at/dms/kjc/backendSupport/CompCommRatio.java
@@ -43,7 +43,7 @@ public class CompCommRatio {
         mults = executionCounts;
         walkSTR(str);
         
-        return ((double)comp)/((double)comm); 
+        return comp/comm; 
     }
 
     // The following structure appears all over the place.  It needs to be abstracted somehow.
