+++ streams/src/at/dms/kjc/common/RawUtil.java
@@ -66,7 +66,7 @@ public class RawUtil extends at.dms.util.Utils {
     
         //prev is a splitter
         double rate = getRRSplitterWeight(prev, node);
-        return ((int)(rate * (double)getCount(counts, prev)));
+        return ((int)(rate * getCount(counts, prev)));
     }
 
     //get the percentage of items sent from splitter prev to  node
